requirements.txt
anyio==3.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
fastapi==0.88.0
h11==0.14.0
httptools==0.5.0
idna==3.4
lxml==4.9.1
pycparser==2.21
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
